requirements.txt
python-dotenv
requests
email-validator
py3dns
discord.py